leverage test, increase some timeouts, join several `it`s
